rule-manager/backend/requirements.txt
Flask
Flask-Cors
psycopg2-binary
google-generativeai
python-dotenv
pydantic
SQLAlchemy
